+++ src/hitrecord.rs
@@ -1,5 +1,6 @@
 use std::rc::Rc;
 use std::marker::{Send, Sync};
+// use std::simd::f32x16;
 use std::sync::Arc;
 use nalgebra::{Vector3, Point3};
 
@@ -8,14 +9,17 @@ use crate::{material::Material, ray::Ray};
 
 #[derive(Clone)]
 pub struct HitRecord {
-    pub t: f64,
-    pub p: Point3<f64>,
-    pub normal: Vector3<f64>,
+    // p + t * normal
+    //when hit with a light ray or photon,the hit method return a ray str
+    pub t: f32,
+    pub p: Point3<f32>,
+    pub normal: Vector3<f32>,
     pub material: Arc<dyn Material>,
 }
 pub trait Hitable : Send + Sync{
-    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord>;
-    fn bounding_box(&self, t0: f64, t1: f64) -> Option<AABB>;
+
+    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord>;
+    fn bounding_box(&self, t0: f32, t1: f32) -> Option<AABB>;
 }
 pub struct UnsafeSyncHitable {
     hitable: Box<dyn Hitable>,
